Added proper size of photo

+++ module1/homework/grayscaleImages/compression.cpp
@@ -7,11 +7,11 @@ std::vector<std::pair<uint8_t, uint8_t>> compressGrayscale(std::array<std::array
 std::array<std::array<uint8_t, width>, height> decompressGrayscale(const std::vector<std::pair<uint8_t, uint8_t>>& compressedMap) {
     std::array<std::array<uint8_t, width>, height> decompressed;
 
-    unsigned int row = 0;
-    unsigned int col = 0;
+    size_t row = 0;
+    size_t col = 0;
 
     for (const auto& [color, count] : compressedMap) {
-        for (unsigned int i = 0; i < count; i++) {
+        for (size_t i = 0; i < count; i++) {
             decompressed[row][col] = color;
             if (col < width - 1) {
                 col++;
